arc-gen/pyproject.toml
[project]
name = "arc-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "airportsdata>=20250909",
    "beautifulsoup4>=4.14.2",
    "fake-useragent>=2.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
